Multithreaded block movement.
- Moved all block animation off of the main thread. Only rotation
(because of respawn) and placement/removal are on the main thread.
- Removed redundant casts.
- Now using time-based animation instead of tick counter based.
- Debug cleanup.
- Added missing closing SQL statements.
- Changed author from Pim to pim16aap2
- Disabled "/bigdoorsenableas" for 1.13.1/2 because it didn't work on
1.13.2. Only 1.12 is supported now.

+++ src/main/java/nl/pim16aap2/bigDoors/Commander.java
@@ -53,7 +53,7 @@ public class Commander
     // Set the availability of the door.
     public void setDoorAvailable(long doorUID)
     {
-        busyDoors.remove((Long) doorUID);
+        busyDoors.remove(doorUID);
     }
 
     // Check if the doors are paused.
